Added tests for can15/oceanic
